"Debuggin endpoint1
endpoint1 total value was wrong, this was due to both number were integer so the result would be truncated,
this issue was solved by transforming any of the two variables into float, this way the result would be a float
that can be ceil to the right number, this action could also be done by simply adding 1 to total. both options are
optimal.

In case per_page was a number outside the range will raise an error.
"